Remove Unnecessary Activation Batching Code
